requirements.txt
fastapi
uvicorn
anthropic
voyageai
chromadb
python-dotenv
